requirements.txt
accelerate==1.9.0.dev0
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.34.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
mysql-connector-python==9.3.0
networkx==3.5
numpy==2.2.4
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pip==25.1.1
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
simplification==0.7.13
six==1.17.0
sympy==1.14.0
tifffile==2025.6.11
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
urllib3==2.5.0
zipp==3.23.0